TASK: Refactor state management to prevent rerendering of the whole layout
UI interactions that triggered a GraphQL query (local or remote) also triggered a full layout refresh which took up
to a 1s on a fast computer. With this change now only about 250ms when updating when changing the GraphQL query for example.
Now all states are managed by hooks and only relevant parts of the UI are therefore updated when interactions happen.
The ones that relate to GraphQL are done via Apollo hooks.
The ones that are only for visual output are done via Recoil.
This way the MediaUiProvider is now much slimer.
Also the naming of some hooks and states has been changed to match the rest of the code