TASK: Make MediaUiProvider aware of NodeCreationDialog
So, the condition to determine whether the selection screen is currently
being shown inside the NodeCreationDialog is a little flimsy right
now...

This commit is just checking, whether the NodeCreationDialog is open,
because we can assume that if `componentDidMount` of
`MediaSelectionScreen` is being called while the dialog is open, the
component definitely must be rendered within the dialog, since the
dialog is modal and thus prevents all other interaction from happening.

Using this solution, we might have more than 1 instance of
`MediaSelectionScreen` at once - one in a SecondaryInspector screen and
one in the NodeCreationDialog. These two do not interfere (verified by
manual testing). This only works because the condition happens to have
the implication described above.

This commmit also prevents the left side bar of the Neos UI from
closing, when the `MediaSelectionScreen` is being rendered within the
NodeCreationDialog.

Additionally, the padding around the `MediaSelectionScreen` is removed,
since the NodeCreationDialog already provides sufficient padding.