BUGFIX: Adjust BottomBar for NodeCreationDialog
This changes the grid layout for the BottomBar as well as its
positioning when rendered in the NodeCreationDialog.

The BottomBar is positioned in a way that makes it look as if it is part
of the dialog.

The grid layout is adjusted, so that each column takes up as much space
as is needed, as opposed to the left-most and right-most columns
reserving 350px worth of space.