BUGFIX: Prevent media ui dialogs from accidentally closing outer Dialogs
This commit creates a (partial) workaround for the issue described in
https://github.com/neos/neos-ui/issues/2925.

It adds a custom decorator component for the Neos UI Dialog that sets
the attribute `data-ignore_click_outside="true"` on the dialog
container.

This solution is crude and relies on a deprecated ReactDOM-API. But for
the time being, it'll at least make the inner dialogs usable in most
cases.

Once https://github.com/neos/neos-ui/issues/2925 is fixed, this solution
should be removed.