FEATURE: Improve visibility of active collection
When a tag is selected it’s not always obvious that
it’s parent collection might also be selected and therefore filter the current view.

With this change the icon is change from ‚folder‘ to ‚folder-open‘ and the active state is set for the parent collection of a selected tag.

Resolves: #159