BUGFIX: Close clipboard when an asset source is selected
refs: #139

When the clipboard screen is open, it is still possible to select an
asset source. But since the clipboard screen remains open, it seems as
if nothing happened upon selecting the asset source.

This commit closes the clipboard screen explicitly when an asset source
is selected via the left sidebar.