BUGFIX: Close clipboard when an asset collection is selected
refs: #139

When the clipboard screen is open, it is still possible to select an
asset collection. But since the clipboard screen remains open, it seems
as if nothing happened upon selecting the asset collection.

This commit closes the clipboard screen explicitly when an asset
collection is selected via the left sidebar.